Add StateSenderV2
