+++ custom_components/ta_coe/state_sender_v1.py
@@ -1,18 +1,11 @@
-"""CoE state sender to send values."""
-from dataclasses import dataclass
+"""CoE state sender to send values V1."""
 from typing import Any
 
 from ta_cmi import CoE, CoEChannel
-from ta_cmi.const import UNITS_EN, ChannelMode
+from ta_cmi.const import ChannelMode
 
-from custom_components.ta_coe.const import _LOGGER, DIGITAL_DOMAINS
-from custom_components.ta_coe.state_sender import StateSender
-
-
-@dataclass
-class AnalogValue:
-    value: float
-    unit: str
+from custom_components.ta_coe.const import _LOGGER
+from custom_components.ta_coe.state_sender import AnalogValue, StateSender
 
 
 class StateSenderV1(StateSender):
@@ -22,23 +15,8 @@ class StateSenderV1(StateSender):
         """Initialize."""
         super().__init__(coe, entity_list)
 
-        self._index_from_id: dict[str, str] = {}
-        self._digital_states: dict[str, bool] = {}
-        self._analog_states: dict[str, AnalogValue] = {}
-
         self._init_digital_states()
         self._init_analog_states()
-        self._init_generate_id_mapping()
-
-    def has_entities(self) -> bool:
-        """Check if the sender has entities in the entity_list"""
-        return len(self._entity_list) > 0
-
-    @staticmethod
-    def _is_domain_digital(entity_id: str) -> bool:
-        """Check if an entity domain is digital."""
-        domain = entity_id[0 : entity_id.find(".")]
-        return domain in DIGITAL_DOMAINS
 
     def _init_digital_states(self) -> None:
         """Create an empty digital states dict."""
@@ -55,24 +33,6 @@ class StateSenderV1(StateSender):
             if not self._is_domain_digital(entity_id):
                 self._analog_states[str(index)] = AnalogValue(0, "0")
                 index += 1
-
-    def _init_generate_id_mapping(self) -> None:
-        """Create the entity index mapping."""
-        digital_index = 0
-        analog_index = 0
-        for entity_id in self._entity_list.values():
-            if self._is_domain_digital(entity_id):
-                self._index_from_id[entity_id] = str(digital_index)
-                digital_index += 1
-            else:
-                self._index_from_id[entity_id] = str(analog_index)
-                analog_index += 1
-
-    def update_digital_manuel(self, entity_id: str, state: bool):
-        """Update a digital state without sending update."""
-        _LOGGER.debug(f"Update digital value without update {entity_id}: {state}")
-        index = self._index_from_id[entity_id]
-        self._digital_states[index] = state
 
     def _build_digital_page(self):
         """Build the digital page."""
@@ -102,26 +62,6 @@ class StateSenderV1(StateSender):
             await self._coe.send_digital_values(page[-16:], True)
         else:
             await self._coe.send_digital_values(page[:16], False)
-
-    def update_analog_manuel(self, entity_id: str, state: float, unit: str):
-        """Update analog state without sending update."""
-        _LOGGER.debug(f"Update analog value without update {entity_id}: {state} {unit}")
-        index = self._index_from_id[entity_id]
-        self._analog_states[index] = AnalogValue(state, unit)
-
-    @staticmethod
-    def _convert_unit_to_id(unit: str) -> str:
-        """Convert the unit to an id."""
-        unit_id: str = "0"
-        for key, value in UNITS_EN.items():
-            if unit == value:
-                unit_id = key
-                break
-
-        if unit_id == "1":
-            unit_id = "46"
-
-        return unit_id
 
     def _build_analog_page(self) -> list[CoEChannel]:
         """Build the analog page."""
