+++ tests/__init__.py
@@ -4,6 +4,8 @@ from typing import Any
 COEAPI_PACKAGE = "ta_cmi.coe_api.CoEAPI.get_coe_data"
 COE_SEND_ANALOG_VALUES_PACKAGE = "ta_cmi.coe.CoE.send_analog_values"
 COE_SEND_DIGITAL_VALUES_PACKAGE = "ta_cmi.coe.CoE.send_digital_values"
+COE_SEND_ANALOG_VALUES_V2_PACKAGE = "ta_cmi.coe.CoE.send_analog_values_v2"
+COE_SEND_DIGITAL_VALUES_V2_PACKAGE = "ta_cmi.coe.CoE.send_digital_values_v2"
 COEAPI_RAW_REQUEST_PACKAGE = "ta_cmi.coe_api.CoEAPI._make_request_get"
 COE_VERSION_CHECK_PACKAGE = "ta_cmi.coe.CoEAPI.get_coe_version"
 COE_CHECK_VERSION_PACKAGE = "ta_cmi.coe.CoE.get_server_version"
@@ -22,6 +24,19 @@ STATE_SENDER_V1_UPDATE_ANALOG_PACKAGE = (
     "custom_components.ta_coe.state_sender_v1.StateSenderV1.update_analog"
 )
 STATE_SENDER_V1_UPDATE = "custom_components.ta_coe.state_sender_v1.StateSenderV1.update"
+STATE_SENDER_V2_UPDATE_DIGITAL_MANUEL_PACKAGE = (
+    "custom_components.ta_coe.state_sender_v2.StateSenderV2.update_digital_manuel"
+)
+STATE_SENDER_V2_UPDATE_ANALOG_MANUEL_PACKAGE = (
+    "custom_components.ta_coe.state_sender_v2.StateSenderV2.update_analog_manuel"
+)
+STATE_SENDER_V2_UPDATE_DIGITAL_PACKAGE = (
+    "custom_components.ta_coe.state_sender_v2.StateSenderV2.update_digital"
+)
+STATE_SENDER_V2_UPDATE_ANALOG_PACKAGE = (
+    "custom_components.ta_coe.state_sender_v2.StateSenderV2.update_analog"
+)
+STATE_SENDER_V2_UPDATE = "custom_components.ta_coe.state_sender_v2.StateSenderV2.update"
 OBSERVER_GET_ALL_STATES = (
     "custom_components.ta_coe.state_observer.StateObserver.get_all_states"
 )
@@ -38,3 +53,13 @@ DUMMY_DEVICE_API_DATA: dict[str, Any] = {
     "last_update_unix": 1680410064.03764,
     "last_update": "2023-04-01T12:00:00",
 }
+
+
+def create_dummy_ids(domain: str, count: int) -> dict[str, Any]:
+    """Create dummy entity id dict."""
+    dummy_data = {}
+
+    for i in range(0, count):
+        dummy_data[str(i)] = f"{domain}.{i}"
+
+    return dummy_data
